fix: move the simulation folder to the services folder
