+++ backend/src/services/simulation/simulation.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { SimulationController } from './simulation.controller';
 import { SimulationService } from './simulation.service';
-import { AdminServiceClient } from '../services/admin-service-client';
+import { AdminServiceClient } from '../admin-service-client';
 
 @Module({
   imports: [HttpModule],
